Migrate Rust numbers test to new framework
