+++ crates/test-rust-wasm/Cargo.toml
@@ -17,10 +17,6 @@ name = "smoke"
 test = false
 
 [[bin]]
-name = "numbers"
-test = false
-
-[[bin]]
 name = "records"
 test = false
