--- crates/test-rust-wasm/src/bin/numbers.rs
@@ -1,3 +1,0 @@
-include!("../../../../tests/runtime/numbers/wasm.rs");
-
-fn main() {}
