+++ tests/runtime-new/numbers/test.wit
@@ -1,0 +1,26 @@
+package test:numbers;
+
+interface numbers {
+  roundtrip-u8: func(a: u8) -> u8;
+  roundtrip-s8: func(a: s8) -> s8;
+  roundtrip-u16: func(a: u16) -> u16;
+  roundtrip-s16: func(a: s16) -> s16;
+  roundtrip-u32: func(a: u32) -> u32;
+  roundtrip-s32: func(a: s32) -> s32;
+  roundtrip-u64: func(a: u64) -> u64;
+  roundtrip-s64: func(a: s64) -> s64;
+  roundtrip-f32: func(a: f32) -> f32;
+  roundtrip-f64: func(a: f64) -> f64;
+  roundtrip-char: func(a: char) -> char;
+
+  set-scalar: func(a: u32);
+  get-scalar: func() -> u32;
+}
+
+world test {
+  export numbers;
+}
+
+world runner {
+  import numbers;
+}
